+++ src/main/resources/META-INF/gradle-plugins/org.shipkit.release-notes.properties
@@ -1,1 +1,1 @@
-implementation-class=org.shipkit.internal.gradle.ReleaseNotesPlugin
+implementation-class=org.shipkit.internal.gradle.notes.ReleaseNotesPlugin
